requirements.txt
pymupdf
pillow
rich